Introduced IO error.
The IO Monad itself now contains the notion of an IO error.
It will propagate to the end of a chain and eventually be passed out
when calling `unsafePerform()`.

I had hoped we wouldn't need this, that those functions that can be
unsafe could return for instance an Either. However, I realized this IO
Either stack is hard to chain, and so the IO errors are now embedded in
the IO monad itself.

Some options to 'catch' IO errors should probably be added.